test: Make Services test helpers stateless
[ upstream commit 91d6364100a605bba85a13209c763b08e3062524 ]

For the sake of readability, in the next commit we are going to move the
helpers to another file. Therefore, we need to free them from any global
var.

For now it's just a dummy move of arguments. In the future commits, some
helper functions with many args will accept a struct.

Signed-off-by: Martynas Pumputis <m@lambda.lt>