workflows: Improve the change check for issue_comment triggers
Joe reported that the CI 3.0 tests were triggered in his pull request,
even though it only touched test/ code and we should skip those tests in
that case [1]. The logs revealed that happened because the workflow had
an incorrect view of the pull request diff (saw more changes than
existed).

To retrieve the set of changed files, we instruct the paths-filter
GitHub action to compare .base.sha and .head.sha. .head.sha points to
the tip of the pull request branch. .base.sha however doesn't point to
the last master commit of the pull request branch (the commit on which
the pull request is based). Instead it points to the latest master
commit at the time you trigger the workflows.

Unfortunately, there doesn't seem to be an easy way to retrieve the
commit on which the pull request is based. Instead, we can keep
.base.sha and compare that to the .merge_commit_sha. The
.merge_commit_sha is computed by GitHub and corresponds to the pull
request's commits merged in .base.sha, so the diff should have exactly
the files the pull request touched.

Of course, GitHub sometimes fails to generate the merge commit because
it can't resolve the conflicts automatically. In such case, we have to
resort to the old approach which might include too many changed files
and trigger CI 3.0 when it's unnecessary. That's what you get for having
outdated pull requests :-p

1 - https://github.com/cilium/cilium/blob/v1.10.2/.github/workflows/conformance-aks.yaml#L69
Reported-by: Joe Stringer <joe@cilium.io>
Signed-off-by: Paul Chaignon <paul@cilium.io>